Add files via upload
In the programming language of your choice, write a program that parses a sentence and replaces each word with the following: first letter, number of distinct characters between first and last character, and last letter.  For example, Smooth would become S3h, as would become a0s, and a would become a0.  Words are separated by spaces or non-alphabetic characters and these separators should be maintained in their original form and location in the answer.